Update parasite_drag_propulsor.py
Propulsor wetted area was computed inside this module. It was inserted a try structure, and the wetted area will be calculated only in cases where propulsor.areas.wetted isn't informed.
The method for form factor estimation is based in a modified finess ratio, considering engine intake and exit areas. This modified finess ratio allows the use of same form factor correlation used for fuselages. 
In the cases where this modified finess ratio can't be calculated the actual finess ratio must be used, but with a different correlation.
The engine parameters used for modified finess ratio estimation looks outdated.